add snapshot tests and extract tweetview
